requirements.txt
# Web development
django==4.2.0
djangorestframework==3.14.0
django-cors-headers==3.14.0

# Data analysis
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
scikit-learn==1.2.2

# Testing
pytest==7.3.1

# Database
psycopg2-binary==2.9.6

# Utilities
python-dotenv==1.0.0# Dependencias para desarrollo y testing
coverage==7.3.0
pytest-django==4.5.2
flake8==6.0.0
black==23.3.0
isort==5.12.0
